Extract preview inventory panel
